投稿一覧の表示を修正
