--- frontend/src/features/posts/index.ts
@@ -1,1 +1,0 @@
-export * from './Posts';
